renames source to sql_database so the section name is correct
