+++ pipelines/sql_database/__init__.py
@@ -1,3 +1,3 @@
-from pipelines.sql_database.source import sql_database
+from pipelines.sql_database.sql_database import sql_database
 
 __all__ = ['sql_database']
